Began working on the type-checking
Formated some source files (Block, Formal, MainClass, Type)
Added scopeable inteface (Block, ClassDeclSimple, If, MainClass, MethodDecl)
Added toString method (Formal, MethodDecl)
Added equals (Identifier, MethodDecl)
Added hashCode (MethodDecl)

Added generics to Visitor (return type for every visit method) and made the
corresponding changes in DepthFirstVisitor and SyntaxTreePrinter, which now
returns Void instead of void.

Updated SymbolTableVisitor with changes in regard to Scopable

TypeMapping now contains its children aswell, in order to be able to check what
type something has (need to traverse the namespace for that)
This also means that it contains a static TypeMapping reference to the
top-scope, i.e the program scope.
I also separated variabledeclarations from methodeclarations (as they should
be) and implemented support for methods with different parameters.
The way support for several diffrent methods with the same name is implemented
is through a hashmap with the name as key and a list as value, this might give
som performance penalty if there are a lot of overloaded members (as the search
is linear in the number of members with the same name). However it was the
cleanest way i could think of to implement it.

have begun implement TypeVisitor. Have had troubles with in which way
everything should be connected, this is what I came up with. visit(Call) are
complex and needs some testing.