Add error handling to lexer and update parser.
Lexer now has a fallback pattern that matches everything that is not a
valid token and throws a NoTokenMatchException if matched. Parser takes
care of this by specifying that next_token() throws that exception and
in scan() catches it and continue to ask for more tokens until a valid
is found (including sym.EOF).